[ADD] 카테고리 생성 기능

+++ frontend/src/containers/Categories/CategoriesContainer.tsx
@@ -2,26 +2,55 @@ import React, { useCallback, useEffect, useState } from "react";
 import { observer } from "mobx-react";
 import Categories from "../../components/Categories";
 import useStore from "../../util/lib/hooks/useStore";
-import { CategoryPostsResponse } from "../../util/types/Response";
+import {
+  CategoriesResponse,
+  CategoryPostsResponse,
+} from "../../util/types/Response";
 import { toast } from "react-toastify";
 import ModalContainer from "../Modal/ModalContainer";
 import Portal from "../../components/common/Portal";
 import CategoryDeleteAlert from "../../components/Categories/CategoryDeleteAlert";
+import isEmpty from "../../util/lib/isEmpty";
+import removeLastBlank from "../../util/lib/removeLastBlank";
 
 const CategoriesContainer = ({}) => {
   const { store } = useStore();
   const {
     categoryPosts,
+    handleCreateCategory,
+    handleCategories,
     handleCategoryPosts,
     handleDeleteCategory,
   } = store.CategoryStore;
   const { admin, login } = store.UserStore;
 
-  const [categoryIdx, setCategoryIdx] = useState<number>(0);
-  const [modifyMode, setModifyMode] = useState<boolean>(false);
-  const [deleteMode, setDeleteMode] = useState<boolean>(false);
   const [isShow, setIsShow] = useState<boolean>(false);
   const [isOpen, setIsOpen] = useState<boolean>(false);
+
+  const [categoryName, setCategoryName] = useState<string>("");
+  const [categoryIdx, setCategoryIdx] = useState<number>(0);
+
+  const [createMode, setCreateMode] = useState<boolean>(false);
+  const [modifyMode, setModifyMode] = useState<boolean>(false);
+  const [deleteMode, setDeleteMode] = useState<boolean>(false);
+
+  const handleCreateCategoryCallback = useCallback(async () => {
+    if (admin && login) {
+      await handleCreateCategory(removeLastBlank(categoryName))
+        .then((res: Response) => {
+          handleCategoryPostsCallback();
+          handleCategoriesCallback();
+          setCreateMode(false);
+        })
+        .catch((err: Error) => {
+          if (err.message.indexOf("409")) {
+            toast.error("중복된 카테고리에요.");
+          } else {
+            toast.error("악! 카테고리 생성에 실패했어요.");
+          }
+        });
+    }
+  }, [categoryName]);
 
   const handleCategoryPostsCallback = useCallback(async () => {
     await handleCategoryPosts()
@@ -31,14 +60,25 @@ const CategoriesContainer = ({}) => {
       });
   }, []);
 
-  const handleDeleteCategoryCallback = useCallback(async () => {
-    await handleDeleteCategory(categoryIdx)
-      .then(() => {
-        handleCategoryPostsCallback();
-      })
+  const handleCategoriesCallback = useCallback(async () => {
+    await handleCategories()
+      .then((res: CategoriesResponse) => {})
       .catch((err: Error) => {
-        toast.error("이런! 카테고리 삭제에 실패했어요.");
+        toast.error("이런! 카테고리 조회에 실패했어요.");
       });
+  }, []);
+
+  const handleDeleteCategoryCallback = useCallback(async () => {
+    if (admin && login) {
+      await handleDeleteCategory(categoryIdx)
+        .then(() => {
+          handleCategoryPostsCallback();
+          handleCategoriesCallback();
+        })
+        .catch((err: Error) => {
+          toast.error("이런! 카테고리 삭제에 실패했어요.");
+        });
+    }
   }, [categoryIdx]);
 
   const showModalCallback = useCallback(() => {
@@ -56,6 +96,18 @@ const CategoriesContainer = ({}) => {
     showModalCallback();
     setCategoryIdx(idx);
   };
+
+  const keyDownListener = useCallback(
+    (e: React.KeyboardEvent<HTMLInputElement>) => {
+      if (
+        !isEmpty(categoryName) &&
+        (e.key === "Enter" || e.key === "NumpadEnter")
+      ) {
+        handleCreateCategoryCallback();
+      }
+    },
+    [categoryName, handleCreateCategoryCallback]
+  );
 
   const deleteClickHandler = useCallback(() => {
     handleDeleteCategoryCallback();
@@ -80,9 +132,16 @@ const CategoriesContainer = ({}) => {
         categoryPosts={categoryPosts}
         admin={admin}
         login={login}
+        categoryName={categoryName}
+        createMode={createMode}
+        modifyMode={modifyMode}
         deleteMode={deleteMode}
+        setCreateMode={setCreateMode}
+        setModifyMode={setModifyMode}
         setDeleteMode={setDeleteMode}
+        setCategoryName={setCategoryName}
         deleteCategoryHandler={deleteCategoryHandler}
+        keyDownListener={keyDownListener}
       />
     </>
   );
